requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.7.0
billiard==3.6.4.0
black==22.10.0
botocore==1.27.96
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==6.5.0
cryptography==38.0.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.3
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-countries==7.4.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-seed==0.3.1
djangorestframework==3.14.0
elasticsearch==7.17.7
elasticsearch-dsl==7.4.0
Faker==15.3.3
filelock==3.8.0
flake8==5.0.4
flake8-isort==5.0.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.4
isort==5.10.1
jmespath==1.0.1
kombu==5.2.4
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.3
prompt-toolkit==3.0.32
psycopg2==2.9.5
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
razorpay==1.3.0
redis==4.3.4
reportlab==3.6.12
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toposort==1.7
twilio==7.15.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
Unidecode==1.1.2
urllib3==1.26.12
vine==5.0.0
virtualenv==20.16.5
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
